요청결과를 serve하기 위해 데이터를 가공, 계산하는 도메인 로직을 도메인 서비스로 이동
