+++ src/applications/lotto/lotto.service.ts
@@ -2,37 +2,18 @@ import { notFound } from '@hapi/boom';
 
 import { Datetime } from '../../domains/common-types/datetime.type';
 
-import { Game } from '../../domains/game/game.model';
 import { Lotto } from '../../domains/lotto/lotto.model';
 import { User } from '../../domains/user/user.model';
 import { LottoNumber } from '../../domains/lotto/lottoNumber.model';
 import { UserId } from '../../domains/user/user.types';
-import { GameId, GameState } from '../../domains/game/game.types';
-import {
-    LottoId,
-    PurchaseId,
-    StateOfNumber,
-} from '../../domains/lotto/lotto.types';
+import { GameId } from '../../domains/game/game.types';
 import { LottoRepository } from '../../infrastructures/lotto/lotto.repository';
 import { UserRepository } from '../../infrastructures/user/user.repository';
 import { getConn } from '../../libs/mysql-client';
-
-type Lottos = Array<{
-    id: LottoId;
-    rank: number;
-    winningAmout: number;
-    lottoNumbers: Array<{
-        number: number;
-        state: StateOfNumber;
-    }>;
-}>;
-
-type PurchasedLottos = Array<{
-    purchaseId: PurchaseId;
-    gameId: GameId;
-    purchasedAt: Datetime;
-    lottos: Lottos;
-}>;
+import {
+    PurchasedLottos,
+    LottoViewPurchaseService,
+} from '../../domains/lotto/services/lotto.view-purchase.service';
 
 interface GetLottosByUserId {
     userId: UserId;
@@ -48,39 +29,9 @@ export class LottoService {
             throw notFound();
         }
 
-        const purchasedLottos: PurchasedLottos = [];
-
-        lottos.forEach(lotto => {
-            const idx = findIndex(
-                purchasedLottos,
-                'purchaseId',
-                lotto.purchaseId,
-            );
-            // purchaseId가 존재하지 않으면 새로 만들고
-            if (idx === -1) {
-                const lottos: Lottos = [];
-                lottos[lotto.order - 1] = {
-                    id: lotto.id,
-                    rank: calculateRank(lotto.lottoNumbers),
-                    winningAmout: lotto.winningAmount,
-                    lottoNumbers: lotto.lottoNumbers,
-                };
-                purchasedLottos.push({
-                    purchaseId: lotto.purchaseId,
-                    gameId: lotto.gameId,
-                    purchasedAt: lotto.purchasedAt,
-                    lottos,
-                });
-            } else {
-                // purchaseId가 존재하면 lottos 데이터만 넣는다.
-                purchasedLottos[idx].lottos[lotto.order - 1] = {
-                    id: lotto.id,
-                    rank: calculateRank(lotto.lottoNumbers),
-                    winningAmout: lotto.winningAmount,
-                    lottoNumbers: lotto.lottoNumbers,
-                };
-            }
-        });
+        const purchasedLottos = new LottoViewPurchaseService().processPurchasedLottoList(
+            lottos,
+        );
 
         return {
             userId,
@@ -161,62 +112,3 @@ export class LottoService {
     // this.lottoRepository.save(lottosNotPulled);
 }
 // }
-
-function calculateRank(lottoNumbers: LottoNumber[]): number {
-    let win = 0;
-    let bonus = 0;
-    let not = 0;
-    let before = 0;
-    lottoNumbers.forEach(lottoNumber => {
-        if (lottoNumber.state === 'WIN') {
-            win += 1;
-        } else if (lottoNumber.state === 'BONUS') {
-            bonus += 1;
-        } else if (lottoNumber.state === 'NOT') {
-            not += 1;
-        } else if (lottoNumber.state === 'BEFORE') {
-            before += 1;
-        } else {
-            // 에러
-        }
-    });
-    if (before === 0) {
-        if (win + bonus + not !== 6) {
-            // 에러
-        }
-    } else {
-        if (before !== 6) {
-            /// 에러
-        }
-    }
-
-    if (bonus > 1) {
-        // 에러
-    }
-
-    if (win === 6) {
-        return 1;
-    }
-    if (win === 5) {
-        if (bonus === 1) {
-            return 2;
-        }
-        return 3;
-    }
-    if (win === 4) {
-        return 4;
-    }
-    if (win === 3) {
-        return 3;
-    }
-    return 0;
-}
-
-function findIndex<T>(list: Array<T>, key: string, value: any): number {
-    for (var i = 0; i < list.length; i++) {
-        if (list[i][key] === value) {
-            return i;
-        }
-    }
-    return -1;
-}
